[v9] fix: Use float mapping for fractional service/agent monitoring fields

+++ src/DIRAC/MonitoringSystem/Client/Types/AgentMonitoring.py
@@ -34,8 +34,8 @@ class AgentMonitoring(BaseType):
                 "Status": {"type": "keyword"},
                 "Location": {"type": "keyword"},
                 "MemoryUsage": {"type": "long"},
-                "CpuPercentage": {"type": "long"},
-                "CycleDuration": {"type": "long"},
+                "CpuPercentage": {"type": "float"},
+                "CycleDuration": {"type": "float"},
             }
         )
